aitemintellex/requirements.txt
asgiref==3.8.1
Django==4.0
django-quill-editor==0.1.40
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1
